Split form component
