+++ resources/views/page_builder/form.blade.php
@@ -1,40 +1,5 @@
-@php($form = Statamic::tag('form:create')->params(['in' => $form->value()->handle(), 'js' => 'vue'])->fetch())
-
-@if($form['success'])
-    <div>@lang('The form was submitted successfully')</div>
-    @return
-@endif
-
-<form-conditions :initial-data='@json(Arr::pluck($form['fields'], 'default', 'handle'))' v-slot="{ formData, submit }">
-    <form @submit.prevent="submit" @attributes($form['attrs'])>
-        @foreach($form['params'] as $name => $value)
-            <input type="hidden" name="{{ $name }}" value="{{ $value }}">
-        @endforeach
-
-        @if($form['honeypot'] ?? false)
-            <input name="{!! $form['honeypot'] !!}" class="hidden" value="">
-        @endif
-
-        <div class="grid grid-cols-12 gap-3">
-            @foreach($form['fields'] as $field)
-                <div
-                    @class([
-                        'col-span-3' => $field['width'] == '25',
-                        'col-span-4' => $field['width'] == '33',
-                        'col-span-6' => $field['width'] == '50',
-                        'col-span-8' => $field['width'] == '66',
-                        'col-span-9' => $field['width'] == '75',
-                        'col-span-12' => $field['width'] == '100',
-                    ])
-                    v-if="{!! $form['show_field'][$field['handle']] ?? true !!}"
-                >
-                    @includeFirstSafe(['form_fields.' . $field['type'], 'rapidez-statamic::form_fields.' . $field['type']], optionalDeep($set)->toArray())
-                </div>
-            @endforeach
-        </div>
-
-        <x-rapidez::button.secondary type="submit">
-            @lang('Submit')
-        </x-rapidez::button.secondary>
-    </form>
-</form-conditions>
+<div class="component form">
+    <div class="container">
+        <x-rapidez-statamic::form :formHandle="$form->value()->handle()" :set="$set" successText="{{ __('The form was submitted successfully') }}" buttonText="{{ __('Submit') }}" />
+    </div>
+</div>
